requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
face_recognition
numpy
pillow
python-dotenv
